requirements.txt
requests
json
pyyaml
discord.py
revChatGPT
GoogleBard